LibVideo: Factor types out from VP9Decoder to their own files
VP9Decoder.h had a bunch of internal classes and types, such as
BitStream and TXSize, which have been moved to their own files for
better readability and management.
