+++ Userland/Libraries/LibVideo/CMakeLists.txt
@@ -1,7 +1,10 @@
 set(SOURCES
     MatroskaDocument.h
     MatroskaReader.cpp
+    VPX/BitStream.cpp
     VPX/Constants.h
+    VPX/Enums.h
+    VPX/LookupTables.h
     VPX/ProbabilityTables.cpp
     VPX/SyntaxElementCounter.cpp
     VPX/VP9Decoder.cpp
